refactor(tasks): extract task api base class
