+++ src/api/APIFactory.js
@@ -14,11 +14,11 @@ import SearchAPI from './Search';
 import RecentsAPI from './Recents';
 import VersionsAPI from './Versions';
 import CommentsAPI from './Comments';
-import TasksAPI from './Tasks';
-import TaskAssignmentsAPI from './TaskAssignments';
-import TasksNewAPI from './TasksNew';
-import TaskCollaboratorsAPI from './TaskCollaborators';
-import TaskLinksAPI from './TaskLinks';
+import TasksAPI from './tasks/Tasks';
+import TaskAssignmentsAPI from './tasks/TaskAssignments';
+import TasksNewAPI from './tasks/TasksNew';
+import TaskCollaboratorsAPI from './tasks/TaskCollaborators';
+import TaskLinksAPI from './tasks/TaskLinks';
 import FileAccessStatsAPI from './FileAccessStats';
 import UsersAPI from './Users';
 import MetadataAPI from './Metadata';
