+++ src/api/__tests__/APIFactory-test.js
@@ -9,10 +9,10 @@ import SearchAPI from '../Search';
 import RecentsAPI from '../Recents';
 import VersionsAPI from '../Versions';
 import CommentsAPI from '../Comments';
-import TasksAPI from '../Tasks';
-import TasksNewAPI from '../TasksNew';
-import TaskCollaboratorsAPI from '../TaskCollaborators';
-import TaskLinksAPI from '../TaskLinks';
+import TasksAPI from '../tasks/Tasks';
+import TasksNewAPI from '../tasks/TasksNew';
+import TaskCollaboratorsAPI from '../tasks/TaskCollaborators';
+import TaskLinksAPI from '../tasks/TaskLinks';
 import FileAccessStatsAPI from '../FileAccessStats';
 import MetadataAPI from '../Metadata';
 import FileCollaboratorsAPI from '../FileCollaborators';
