+++ src/api/tasks/TaskAssignments.js
@@ -4,34 +4,15 @@
  * @author Box
  */
 
-import Base from './Base';
+import TasksBase from './TasksBase';
 import {
     PERMISSION_CAN_COMMENT,
     ERROR_CODE_CREATE_TASK_ASSIGNMENT,
     ERROR_CODE_UPDATE_TASK_ASSIGNMENT,
     ERROR_CODE_DELETE_TASK_ASSIGNMENT,
-    HTTP_STATUS_CODE_RATE_LIMIT,
-    HTTP_STATUS_CODE_INTERNAL_SERVER_ERROR,
-    HTTP_STATUS_CODE_NOT_IMPLEMENTED,
-    HTTP_STATUS_CODE_BAD_GATEWAY,
-    HTTP_STATUS_CODE_SERVICE_UNAVAILABLE,
-    HTTP_STATUS_CODE_GATEWAY_TIMEOUT,
-} from '../constants';
+} from '../../constants';
 
-const RETRYABLE = [
-    HTTP_STATUS_CODE_RATE_LIMIT,
-    HTTP_STATUS_CODE_INTERNAL_SERVER_ERROR,
-    HTTP_STATUS_CODE_NOT_IMPLEMENTED,
-    HTTP_STATUS_CODE_BAD_GATEWAY,
-    HTTP_STATUS_CODE_SERVICE_UNAVAILABLE,
-    HTTP_STATUS_CODE_GATEWAY_TIMEOUT,
-];
-
-class TaskAssignments extends Base {
-    constructor({ retryableStatusCodes = RETRYABLE, ...options }: Options) {
-        super({ ...options, retryableStatusCodes });
-    }
-
+class TaskAssignments extends TasksBase {
     /**
      * API URL for task assignments. Getting a list of assignments "/tasks/id/assignments" does not give us the fields
      * we need. So instead we will only perform GET operations on an assignment by assignment basis,
