+++ src/api/tasks/__tests__/TaskCollaborators-test.js
@@ -1,4 +1,4 @@
-import { API_PAGE_LIMIT } from '../../constants';
+import { API_PAGE_LIMIT } from '../../../constants';
 import TaskCollaborators from '../TaskCollaborators';
 
 let taskCollaborators;
