+++ src/api/tasks/__tests__/Tasks-test.js
@@ -1,5 +1,5 @@
 import Tasks from '../Tasks';
-import { PERMISSION_CAN_COMMENT } from '../../constants';
+import { PERMISSION_CAN_COMMENT } from '../../../constants';
 
 let tasks;
